+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+4.1.0 (2014-08-18)
+------------------
+
+  * implemented issue #65
+    * added `stubbles\input\web\WebRequest::userAgent()`
+    * deprecated user agent injection using `stubbles\input\web\useragent\UserAgentProvider`, will be removed with 5.0.0
+
+
 4.0.0 (2014-08-17)
 ------------------